requirements.txt
blinker==1.7.0
bottle==0.12.25
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.44.3
fsspec==2023.10.0
idna==3.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
proxy-tools==0.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pythonnet==3.0.3
pytz==2023.3.post1
pywebview==4.4.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
tk==0.1.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.210
urllib3==2.1.0
Werkzeug==3.0.1
